Working Changes
These changes contain a lot of corrections to function and variable
names to start with uppercase.

Scenes now only perform renders and updates to the game object. The
GameObject will update the transition on each render. This prevents
weird hitching effects since the transition may run faster on another
thread, hence performing the transition before drawing it makes more
sense. Note that this can slow down rendering performance with lots of
transitions. GameObjects refer to their parent (Scene), this allows the
game object SendToTop and SendToBottom zIndex functions to function, and
this will call Scene's RefreshZIndexes which sorts the GameObject
queue. This worsens coupling between these 2 classes which is inevitable
right now. By doing this, we won't have to sort the GameObjects
collection on each render which should improve performance. Small
improvements to return early have been made for rendering.

Transitions now wait to create a start time during the first value
rather than when it is enabled, this should help prevent transitions
from starting too early.

The Raindrop demo has been improved and simplified. This demo is good
for testing the GameObject collection in the scene and the transitions
collection in the GameObject. Transitions have been improved to use a
dictionary to ease add/removal of the TransitionPropertyItem which is
created at runtime. The ID can be generated through
TransitionPropertyItem.GetId() statically or per instance. The
GameObject collection is a bit more difficult and I have experimented
with a 2 pooling classes. Both poll in between access. From
experimenting I think the linked list is more efficient due to the
simplicity of performing removals during each transition per node and
then adding nodes at the end, but sorting is difficult and not possible
(yet). The ModifiableList just uses a pool of cached calls (e.g.
Add/Remove/InsertAt) which allows sorting by replacing the whole list
but I believe it's slower. Both are quite complicated since the pools
can only run at specific times. I believe if sorting can be achieved the
linked list will be best as swapping the links between nodes is
preferrable.

The Sine Demo has some minor improvements with SineCircle ToString
formatting. Pausing was using the Game pause function which is not
appropriate. Instead a virtual pause menu makes sense as would most
games, as this method pauses transitions and keeps everything rendered.
This allows for the "Paused" text to be shown and animated. The pause
method should only be used to save rendering during an inactive window,
rather than a pause menu/screen.

The window in GameContainer can transition to fullscreen on any screen
it intersects unlike before which used the primary screen. The KeyPress
event has been disabled as the behaviour is not useful for us.