Finish working changes for weekend
These additionally fix some small issues discovered afterwards.

Transitions are updated before a render still but they must be done
separately since forcing transitions to be updated through GameObject
inheritance is not a good idea (as seen in SineCircle). This small bug
allowed me to add a SineCircle ShowMidPoint boolean which is quite nice
debugging information and can be toggled in the demo pressing F5.

The ModifiableList is still not in a desirable state as it still causes
issues whilst being modified so the pool has been moved to a LinkedList
to act as a stack with the pool utilised all the time. This seems to be
relatively stable now.

This is the end of some fun over the weekend.